+++ CHANGELOG.md
@@ -3,13 +3,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.0.11] - 2026-09-16
+## [3.0.11] - 2026-09-24
 
 ### Added
 
 - Added optional wave progress bar
 - Ctrl+F now selects all text in the search box for quick replacement
 - Added optional Stop button (Hidden by default, but can be enabled in the settings.)
+- Scroll position is now remembered when switching screens
+- Added buttons to the playback queue to clear the queue and to add the queued songs to a playlist
+- Added album release date to album image tooltips and optionally on the nog playing page
 
 ### Changed